assignment/requirements.txt
pdfplumber
pandas
openpyxl
python-dateutil
regex
pytesseract
pdf2image
pillow
streamlit
